Check for 'compile-and-run` devices (#229)
`aqt_keysight_qpu` is `compile-only`, while `provider.get_backends()`
lists backends the user can run on. I updated the test accordingly.

This once again calls into question if we should:
- (a) include `compile-only` devices
- (b) do (a) but differentiate between the two in the output (i.e., a
dictionary with `compile-only` and `compile-and-run` keys)
- (c) leave it as is (after this PR merge)

In the past we avoided (b) because this would break with what `qiskit`
users would expect.