Fail when requesting non-existent backends (#243)
This fixes #230 by just keeping a list of valid device prefixes around
(there might be objection to this). This is similar to how CSS checks
for a specific prefix except I added a method to validate the prefix
when you instantiate the provider (so it should fail immediately rather
than when a user goes to make a request to a compiler).