Add `oxford` prefix and remove D-Wave (#272)
We can't run on Lucy right now, but there's a server-side issue that
will fix this (i.e., remove backends that we can't yet run on, which is
just Lucy AFAIK). For now to get the deploy in we'll just add the prefix
since it's listed in `SuperstaQProvider.backends()`.

We'll also remove D-Wave since AWS took it off Braket and it's also
needed to fix the integration test / SuperstaQ deploy. (There's also a
server-side issue addressing this.)

Integration test passed on `prefix-and-dwave`:

![image](https://user-images.githubusercontent.com/18367737/202729904-c5f5c4f7-c06f-49db-a4ff-68a0fbb0eb0f.png)