Update dependencies to match SuperstaQ (#70)
Quick fix to keep SupermarQ dependencies up-to-date with SuperstaQ. 

@vtomole I saw another PR #62 and some others in SuperstaQ that you have
an automated way to update dependencies. Could we use that here to keep
SupermarQ up to date with the latest css and qss deployments?

The css and qss packages are mainly used for circuit submissions so
automatically updating them shouldn't introduce breaking changes, plus
we should probably keep SupermarQ up-to-date with them regardless.