fix requirements check and sync versions (#390)
now that we sync package versions and publish everything at once, we
should be putting the same version number in every package's
requirements.txt. this fails the requirements check though, because we
have to update the requirements before publishing anything to pypi

this is a hotfix which uses `gss.__version__` instead of the pypi
version if the former is larger. we should really rethink how the
requirements check should work now though