reorganized docs superstaq and deleted unnecessary application examples (#587)
reorganizing the files in our apps directory; we got rid of the other work subdir and moved supercheq into the applications subdir. We also removed all the applications files in the applications subdir except for the max sharpe ratio. in the next PR we will edit max sharpe ratio accordingly