Upper bound notebook `pytest` threads to 4 (#1244)
Various notebook check timeouts/failures, like in
https://github.com/Infleqtion/client-superstaq/issues/1129, may
potentially be attributed to spawning too many kernels in a short amount
of time, trying to use a port already in use. This PR hence limits the
number of threads to a maximum of four in an effort to reduce the
likelihood of such scenarios from happening.

As of this PR, the number of workers currently spawned by each notebook
check is:

- `cirq-superstaq`: 9 workers
- `qiskit-superstaq:` 8 workers
- `supermarq-benchmarks`: 6 workers

Signed-off-by: Bharath <bharath.thotakura@infleqtion.com>