Support `'ibm_quantum_platform'` and deprecate older channels (#1220)
The IBM Quantum channel is going to be [sunset on July 1st,
2025](https://docs.quantum.ibm.com/guides/setup-channel#:~:text=IBM%20Quantum%20Platform%20is%20moving%20and%20this%20version%20will%20be%20sunset%20on%20July%201.%20To%20get%20started%20on%20the%20new%20platform%2C%20read%20the%20migration%20guide.).
This PR thus drops support for it (with a `ValueError` message) and also
adds support for the new `'ibm_quantum_platform'` that is in 'early
access mode' and will presumably be de-facto ibmq channel (per the
release notes of `qiskit-ibm-runtime==0.40.0`):
> This new channel connects to the new IBM Quantum Platform API and is
intended to replace ibm_cloud. In the meantime, the ibm_cloud channel
will redirect to the new API, but its continued use is discouraged.

_Note: intended only to be deployed after July 1st, 2025_

---------

Signed-off-by: Bharath <bharath.thotakura@infleqtion.com>